--- api/src/main/java/com/tersesystems/echopraxia/DefaultFieldBuilder.java
@@ -1,12 +1,0 @@
-package com.tersesystems.echopraxia;
-
-class DefaultFieldBuilder implements Field.Builder {
-
-  private static final Field.Builder instance = new DefaultFieldBuilder();
-
-  private DefaultFieldBuilder() {}
-
-  static Field.Builder singleton() {
-    return instance;
-  }
-}
